1/20 22:09

+++ Coffee_iOS/Coffee_iOS/VC/반제품/HalfProductSecondVC.h
@@ -14,7 +14,7 @@
     NSArray *datas;
     NSArray *datas2;
     
-    int mPosition;
+    NSInteger mPosition;
     NSString *mSample_idx;
     
     NSDictionary *tableDic;
